Add outline of query structure/API.
The API is still missing pretty much any 'real' features, however, it sketches what we want the access layer to look like. We refrain us from adding more functionality for now, as it seems better to do this 'on demand' as soon as we know how we want to query and update the data.