WIP: Items returned by DB are decoupled from the internal DB representation.
This means, that we return logical items holding invariants instead of including DB internals. The most notable change here is moving the creation time from the metadata field (which makes sense in the internal DB representation) to an own logical field (which makes sense from a logical data structure point of view.
What is still missing, is the differentiation between 'last-mod-time' and 'max-mod-time' for folders. We will tackle this with the next refactoring step.