Remove un-needed indices and 'bundle' scanning into one transaction.
The un-needed indices result from the fact that SQLite will create implicit indices on Unique contstrained columns. This is a plain improvement (less storage, same performance). The 'bundling' of DB writes during an initial scan process is less clear cut for performance, we will have to do some better testing in the future to see if the small bump in complexity is worth it.