Simplify structure of scan algorithm.
We pulled together the two code paths of indexing files and directories. It became apparent that much code was duplicated, thus making sense to morge them before adding any more functionality. Additionally, we reverted the 'bundling' into transactions on scans. We re-measured it and see no performance difference on read-heavy workloads. Scan-heavy workloads will have to be re-considered for performance anyways with probably a multithreaded model.