WIP: Handle ignored items in the DB.
We want to ignore items in the DB to allow the sync algorithm to handle partial data stores. We perform the first step for this by introducing the invariant in the DB and handling it during the scan. WIP, as we still need to patch the sync algorithm to account for ignored files.