WIP: Start to add proper sync conflict resolution.
We added a callback method that is invoked on any conflicts during a sync. It then decides how to handle the conflict (choose local, choose remote, do nothing). The basic mechanism works and we started writing tests for some cases. However, it seems like some code passages are redundant, so we will look into solving this (thus hopefull simplifying the rather complex special cases in the sync algorithm). WIP is removed when we implement broader test coverage.