Add first tested version of conflict resolution.
We implement basic tests for all conflict enums, as well as the three test cases described in Russes paper. All interactions work as planned. We therefore declare this a first version of conflict resolution. We would like to add more tests involving ignored files AND conflict resolution in the future. Also, we defenitely require some randomized tests to really rely on the sync algorithm.