Rework inclusion/ignore rules: rules are always consistent with DB items.
Before, the ignore rules (for excluding files from the sync/being indexed) where only loosely coupled to the actual DB content, i.e. it can happen that a file is indexed in the DB that would match an ignore rule (and the other way around). This was fine with the old semantic, however, it can become problematic when we want to introduce more sophisticated ignore rules to enable partial syncs and data stores that transfer changes to other data stores. The new rules are always consistent with the DB (when an ignore rule matches a path, it is also ignored in the DB). To do this, changing ignore rules will always also update the database to match the changes.