WIP: 'transfer' stores, that can carry data.
Transfer stores can be used to carry changes between stores. They detect if any of the stores they know need an item (as it was updated) and keep it until every store has gotten the item. This can be used to e.g. carry changes on a laptop with you until you plug in every hard drive.