Fix: Sync with ignored files skipped child entries.
A short circuiting evaluation of 'flag = flag && recursive_call()' was causing the recursive child calls to not happen after the flag was once false. However, the algorithm should continuing processing children and just remeber that at least one child failed. Added a regression test to show the issue and fix it.